Added new Amazon class for related functionality.
