+++ src/WooCommerce.php
@@ -1257,34 +1257,5 @@ class WooCommerce {
     }
   }
 
-  /**
-   * Replaces the order ID of imported Amazon orders with the custom order ID.
-   * 
-   * @implements woocommerce_amazon_pa_update_checkout_session_payload
-  */
-  public static function woocommerce_amazon_pa_update_checkout_session_payload ($payload, $checkout_session_id, $order) {
-
-    if (empty($payload) || empty($order) || !self::isAmazonPayV2()) {
-  
-      return $payload;
-    }
-
-    $payload['merchantMetadata']['merchantReferenceId'] = $order->get_order_number();
-
-    return $payload;
-  }
-
-  /**
-   * Check if we are using Amazon V2.
-   *
-   * @return bool
-   */
-  public static function isAmazonPayV2(): bool {
-    $is_amazon_pay_active = is_plugin_active('woocommerce-gateway-amazon-payments-advanced/woocommerce-gateway-amazon-payments-advanced.php');
-    if ($is_amazon_pay_active) {
-      return defined('WC_AMAZON_PAY_VERSION') && version_compare(WC_AMAZON_PAY_VERSION, '2.0', '>=');
-    }
-    return false;
-  }
 
 }
